+++ R/expandLenFreq.R
@@ -2,7 +2,7 @@
 #'
 #' @description Creates a vector of lengths for the individuals not measured based on the lengths measured in a subsample of individuals.
 #'
-#' @details Creates a vector of lengths for the individuals not measured based on the lengths measured in a subsample of individuals.  Length categories are created first that begin with the value in \code{startcat} (or the minimum observed value by default) and continue by values of \code{w} until a category value greater than the largest observed length in \code{x}.  Categories of different widths are not allowed.  
+#' @details Creates a vector of lengths for the individuals not measured based on the lengths measured in a subsample of individuals.  Length categories are created first that begin with the value in \code{startcat} (or the minimum observed value by default) and continue by values of \code{w} until a category value greater than the largest observed length in \code{x}.  Categories of different widths are not allowed.
 #'
 #' The resulting \dQuote{expanded} lengths are created by allocating individuals to each length class based on the proportion of measured individuals in the subsample in that length class.  Individuals within a length class are then assigned a specific length within that length class based on a uniform distribution.  Because the expanded number of individuals in a length class is rounded down based on the measured number per length class, not all individuals will initially be assigned a length value.  The remaining individuals are assigned to a length class randomly according to weights based on the proportion of individuals in the measured length classes.  Finally, these individuals are randomly assigned a specific length within the respective length class from a uniform distribution, same as above.
 #'
@@ -76,7 +76,7 @@ expandLenFreq <- function(x,w,additional,
   num <- total-length(x)
   # find the length frequency of measured fish
   lcat <- lencat(x,w=w,startcat=startcat,...)
-  lenfreq <- prop.table(table(lcat))
+  lenfreq <- prop.table(table(lcat,dnn=NULL))
   # length frequency categories (lower limit of bin)
   cats <- as.numeric(rownames(lenfreq))
   # number of expanded individuals per length category
@@ -93,16 +93,16 @@ expandLenFreq <- function(x,w,additional,
   new.lens <- round(new.lens,decimals)
   # if asked, print some summary values of what happened
   if (show.summary) {
-    message("Length Frequency Expansion using:\n",
-            "Measured length frequency of ",length(x)," individuals:")
-    print(lenfreq)
-    message("\nNon-random allocations of ",length(nrand.lens)," individuals by length category.")
-    print(reps)
-    message("\nRandom allocations of ",length(rand.lens)," individuals\n",
-            "With final length frequency table of:")
+    cat("Length Frequency Expansion using:\n","Measured length frequency of",length(x),"individuals:\n")
+    print(round(lenfreq,4))
+    cat("\nNon-random allocations of",length(nrand.lens),"individuals by length category.\n")
+    tmp <- reps
+    names(tmp) <- NULL
+    print(tmp)
+    cat("\nRandom allocations of",length(rand.lens),"individuals\n","With final length frequency table of:\n")
     final.lens <- lencat(new.lens,w=w,startcat=startcat,...)
-    print(table(final.lens))
+    print(table(final.lens,dnn=NULL))
   }
   # return sorted vector of newly assigned lengths
-  sort(new.lens)
+  invisible(sort(new.lens))
 }