+++ tests/testthat/test_expandLenFreq_MSG.R
@@ -11,5 +11,5 @@ test_that("expandLenFreq() messages",{
   expect_error(expandLenFreq(x,1,17,startcat=-2),"must be positive")
   expect_error(expandLenFreq(x,1,total=17),"Total number to expand")
   ## do you get a message
-  expect_message(expandLenFreq(x,1,17),"Length Frequency Expansion using")
+  expect_output(expandLenFreq(x,1,17),"Length Frequency Expansion using")
 })