+++ tests/testthat/test_extraTests_OUT.R
@@ -12,8 +12,10 @@ test_that("extraSS() and lrt() output",{
   expect_true(is.matrix(tmp2))
   expect_equal(nrow(tmp2),2)
   # print() returns x
-  expect_equal(tmp1,print.extraTest(tmp1))
-  expect_equal(tmp2,print.extraTest(tmp2))
+  junk <- capture.output( tmp <- print.extraTest(tmp1))
+  expect_equal(tmp1,tmp)
+  junk <- capture.output( tmp <- print.extraTest(tmp2))
+  expect_equal(tmp2,tmp)
   # lrt() returns a matrix of class extraTest
   tmp1 <- lrt(lm.0,com=lm.1)
   expect_is(tmp1,"extraTest")
@@ -24,6 +26,8 @@ test_that("extraSS() and lrt() output",{
   expect_true(is.matrix(tmp2))
   expect_equal(nrow(tmp2),2)
   # print() returns x
-  expect_equal(tmp1,print.extraTest(tmp1))
-  expect_equal(tmp2,print.extraTest(tmp2))
+  junk <- capture.output( tmp <- print.extraTest(tmp1))
+  expect_equal(tmp1,tmp)
+  junk <- capture.output( tmp <- print.extraTest(tmp2))
+  expect_equal(tmp2,tmp)
 })